orchestrator/pyproject.toml
[project]
name = "jcsandbox-orchestrator"
version = "0.1.0"
description = "JcSandbox — Bot 军团编排控制器，基于 OpenSandbox 管理 jcdo 实例生命周期"
requires-python = ">=3.11"
dependencies = [
    "opensandbox",
    "httpx",
    "pydantic>=2.0",
    "pyyaml",
    "typer>=0.9",
    "requests",
]

[project.scripts]
jcsandbox = "orchestrator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
